+++ dev_server/server.properties
@@ -1,5 +1,5 @@
 #Minecraft server properties
-#Fri Aug 01 10:15:13 CEST 2025
+#Fri Aug 01 10:24:40 CEST 2025
 accepts-transfers=false
 allow-flight=false
 allow-nether=false
