+++ src/main/java/com/beepsterr/plugins/Depends/PlaceholderAPI.java
@@ -1,0 +1,11 @@
+package com.beepsterr.plugins.Depends;
+
+import org.bukkit.entity.Player;
+
+public class PlaceholderAPI {
+
+    public static String setPlaceholders(Player player, String text) {
+        return me.clip.placeholderapi.PlaceholderAPI.setPlaceholders(player, text);
+    }
+
+}
