+++ src/main/java/com/beepsterr/plugins/Library/Types/PlaceholderConditionType.java
@@ -1,0 +1,13 @@
+package com.beepsterr.plugins.Library.Types;
+
+public enum PlaceholderConditionType {
+    EQUALS,
+    NOT_EQUALS,
+    CONTAINS,
+    STARTS_WITH,
+    ENDS_WITH,
+    GREATER_THAN,
+    LESS_THAN,
+    GREATER_THAN_OR_EQUALS,
+    LESS_THAN_OR_EQUALS
+}
