--- src/main/java/dev/beeps/plugins/Depends/PlaceholderAPI.java
@@ -1,11 +1,0 @@
-package dev.beeps.plugins.Depends;
-
-import org.bukkit.entity.Player;
-
-public class PlaceholderAPI {
-
-    public static String setPlaceholders(Player player, String text) {
-        return me.clip.placeholderapi.PlaceholderAPI.setPlaceholders(player, text);
-    }
-
-}
