--- src/main/java/dev/beeps/plugins/Library/Types/PlaceholderConditionType.java
@@ -1,13 +1,0 @@
-package dev.beeps.plugins.Library.Types;
-
-public enum PlaceholderConditionType {
-    EQUALS,
-    NOT_EQUALS,
-    CONTAINS,
-    STARTS_WITH,
-    ENDS_WITH,
-    GREATER_THAN,
-    LESS_THAN,
-    GREATER_THAN_OR_EQUALS,
-    LESS_THAN_OR_EQUALS
-}
